final commit v1.0

+++ src/main/java/br/com/itexto/desafiobackend/repository/AutorRepository.java
@@ -3,5 +3,5 @@ package br.com.itexto.desafiobackend.repository;
 import br.com.itexto.desafiobackend.model.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutorRepository extends JpaRepository<Autor,Integer> {
+public interface AutorRepository extends JpaRepository<Autor,Long> {
 }
